+++ packages/utils/src/math.ts
@@ -1,0 +1,3 @@
+export function isNumber(n: any) {
+  return typeof n === 'number';
+}
